+++ workspace/web_edu/web_edu_lib/src/core/mod.rs
@@ -3,4 +3,6 @@ pub mod product;
 pub mod sale;
 pub mod product_delete;
 pub mod product_variant;
+pub mod product_create;
+pub mod product_variant_create;
 pub mod product_variant_edit;
